Update
-Implemented Zhuoran's reference compression algorithm as a memory/speed tradeoff for locations with high memory requirement
-Optimized the custom BigNum implementation, which decreases the reference calculation time in shallow areas
-Optimized the mpfr/mpir reference calculations by reducing the number of JNA calls
-Implemented Zhuoran's multithreaded BLA table generation
code. (It mostly affects areas with high iteration count, when using compression)
-Modified the drawing algorithms to include more tiling options
-Added convergent bailout as an option
-The zoom sequence renderer of Image Expander will now save the render settings and also the zoom factor on every step
-Added an aspect ratio option in zoom sequence renderer as a temporary workaround, until I implement non-square image format
-Bug Fixes and general improvements

+++ src/fractalzoomer/utils/MathUtils.java
@@ -31,7 +31,7 @@ public class MathUtils {
 
 	public static double fract(double x) {
 		x = Math.abs(x);
-		return x - (int)x;
+		return x - (long)x;
 	}
     
     public static double angleBetween2PointsDegrees(double x1, double y1, double x2, double y2) {
